Fixed the stage2 pe loading code for 64 bit

+++ librt/include/os/osdefs.h
@@ -69,10 +69,10 @@ typedef reg64_t                     reg_t;
  * these are usually fixed no matter arch and include stuff
  * as threading, processing etc */
 typedef unsigned int                IntStatus_t;
-typedef unsigned int                UUId_t; 
+typedef size_t                      UUId_t; 
 typedef unsigned int                Flags_t;
 typedef unsigned                    DevInfo_t;
-typedef void *                      Handle_t;
+typedef void*                       Handle_t;
 
 /* Define some special UUId_t constants 
  * Especially a constant for invalid */
